% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_callbacks.R
\name{callback_lr_init}
\alias{callback_lr_init}
\title{This function should be used with keras callback_lambda()
Use for metrics on each epoch, iteration, batch
When training a model, each epoch has 'n' iterations, where each iteration has 'k' batches}
\usage{
callback_lr_init(logs)
}
\arguments{
\item{logs}{metrics contained in a list generated by R6 class}
}
\value{
iteration information relating to learning rate
}
\description{
This will clear the learning rate history and iteration hist
}
