% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_callbacks.R
\name{callback_lr_log}
\alias{callback_lr_log}
\title{This function should be used with keras callback_lambda
Use for metrics on each epoch, iteration, batch
When training a model, each epoch has 'n' iterations, where each iteration has 'k' batches}
\usage{
callback_lr_log(batch, logs)
}
\arguments{
\item{batch}{created from keras callbacks, logs 'which batch'}

\item{logs}{metrics contained in a list generated by R6 class}
}
\value{
lists for lr_hist and iter_hist
}
\description{
This function logs the learning rate value and iteration number for _hist in global enviroment
}
