% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_callbacks.R
\name{create_lr_rate_parameters_exp}
\alias{create_lr_rate_parameters_exp}
\title{This function generates parameters for learning rate callbacks defined by user input
of number of epochs, the max learning rate willing, tensor used as input for ML or DL model, growth constant, and batch_size}
\usage{
create_lr_rate_parameters_exp(
  epochz,
  max_lr,
  training_data,
  growth_constant,
  process_batch_size
)
}
\arguments{
\item{epochz}{number of epochs the user plans to set for model they wish to train data on}

\item{max_lr}{maximum allowed value for learning rate}

\item{training_data}{A tensor of specific dimension}

\item{growth_constant}{growth_constant for modeling exponential increase}

\item{process_batch_size}{number of batches set when training model}
}
\value{
l_rate and n_iter
}
\description{
This function generates parameters for learning rate callbacks defined by user input
of number of epochs, the max learning rate willing, tensor used as input for ML or DL model, growth constant, and batch_size
}
