% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_create_CNNs.R
\name{create_residual_3dCNN}
\alias{create_residual_3dCNN}
\title{Create CNN architecture}
\usage{
create_residual_3dCNN(Learning_rate = 0.1, decay = FALSE)
}
\arguments{
\item{Learning_rate}{learning rate speed}

\item{decay}{weighted decay rate}
}
\value{
CNN model as returned from keras
}
\description{
Create CNN architecture
}
