% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_crop_LEAs.R
\name{x3p_crop_LEAs}
\alias{x3p_crop_LEAs}
\title{Crop an x3p object}
\usage{
x3p_crop_LEAs(df, obtain_index = TRUE, value, multiple = FALSE, select_col)
}
\arguments{
\item{df}{object of class 'x3p' or a tibble style dataframe containing 'x3p' objects}

\item{obtain_index}{Grid locations from where the crop occured}

\item{value}{window size}

\item{multiple}{clarify whether user is working with a single 'x3p' object or multiple in a tibble style dataframe}

\item{select_col}{if multiple = TRUE, location of the x3p objects}
}
\value{
A list containig cropped x3ps and original indexes if specified
}
\description{
This function allows the user to crop an 'x3p' object into the maximum number of smaller 'x3p' objects
by specifying a window size
}
