% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_get_labels.R
\name{x3p_get_labels}
\alias{x3p_get_labels}
\title{Obtain all labels and majority label from an 'x3p' object's mask}
\usage{
x3p_get_labels(df, decision_ratio, multiple = FALSE, select_col)
}
\arguments{
\item{df}{object of class 'x3p' or a tibble style dataframe containing 'x3p' objects}

\item{decision_ratio}{numeric value specified by the user for majority label}

\item{multiple}{clarify whether user is working with a single 'x3p' object or multiple in a tibble style dataframe}

\item{select_col}{if multiple = TRUE, location of the x3p objects}
}
\value{
A list containing labels and majority labels associated
}
\description{
This function gives the user each label associated with a single or multiple 'x3p' objects
Majority label is defined by a decision ratio specified by the user
}
