% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_interpolate_NA.R
\name{x3p_interpolate_NA}
\alias{x3p_interpolate_NA}
\title{Interpolate Missing Values in surface.matrix of an 'x3p' object}
\usage{
x3p_interpolate_NA(df, multiple = FALSE, select_col)
}
\arguments{
\item{df}{object of class 'x3p' or a tibble style dataframe containing 'x3p' objects}

\item{multiple}{clarify whether user is working with a single 'x3p' object or multiple in a tibble style dataframe}

\item{select_col}{if multiple = TRUE, location of the x3p objects}
}
\value{
A 'x3p' object with a surface.matrix containing no missing values
}
\description{
This function interpolates the missing values contained in a single 'x3p' object or multiple 'x3p' objects
A missing ith jth cell is interpolated by taking the average of that cell's 8 neighbors
}
