% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_labels_to_numeric.R
\name{x3p_labels_to_numeric}
\alias{x3p_labels_to_numeric}
\title{Convert labels to numeric to be used for keras one-hot encoding for DL multiclassification}
\usage{
x3p_labels_to_numeric(df)
}
\arguments{
\item{df}{object of class 'x3p' or a tibble style dataframe containing 'x3p' objects}
}
\value{
A numeric vector
}
\description{
Convert labels to numeric to be used for keras one-hot encoding for DL multiclassification
}
