% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_surf_stat_rescale.R
\name{x3p_lea_df}
\alias{x3p_lea_df}
\title{This function generates even more stats user for heuristic statistics and generating plots}
\usage{
x3p_lea_df(df)
}
\arguments{
\item{df}{tibble style dataframe containing 'x3p' objects}
}
\description{
This function generates even more stats user for heuristic statistics and generating plots
}
