% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_raster_relabel.R
\name{x3p_raster_relabel_boundary}
\alias{x3p_raster_relabel_boundary}
\title{Relabel mask of an 'x3p' object with x and y border specifications
This function allows a user to relabel a mask of an 'x3p' object
The user is required to specify minimum and maximum, x and y locations, beaware of surface dimensions}
\usage{
x3p_raster_relabel_boundary(
  df,
  mask_1,
  mask_2,
  x1,
  x2,
  y1,
  y2,
  multiple = FALSE,
  select_col
)
}
\arguments{
\item{df}{object of class 'x3p' or a tibble style dataframe containing 'x3p' objects}

\item{mask_1}{old mask label the user wishes to replace, must be a color code #8digits}

\item{mask_2}{new mask label user wishes to replace with old mask label, must be a color code #8digits}

\item{x1}{minimum x location...}

\item{x2}{maximum x location...}

\item{y1}{minimum y location...}

\item{y2}{maximum y location...}

\item{multiple}{clarify whether user is working with a single 'x3p' object or multiple in a tibble style dataframe}

\item{select_col}{if multiple = TRUE, column name containing the x3p objects}
}
\value{
'x3p' object with altered mask
}
\description{
Relabel mask of an 'x3p' object with x and y border specifications
This function allows a user to relabel a mask of an 'x3p' object
The user is required to specify minimum and maximum, x and y locations, beaware of surface dimensions
}
