% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_surf_stat_rescale.R
\name{x3p_surface_rescale}
\alias{x3p_surface_rescale}
\title{Get Surface Summary Statistics from x3p object
Rescale Surface of an x3p object}
\usage{
x3p_surface_rescale(
  df,
  obtain = c("minimum", "maximum", "both", "rescale"),
  multiple = FALSE,
  select_col,
  value
)
}
\arguments{
\item{df}{object of class 'x3p' or a tibble style dataframe containing 'x3p' objects}

\item{obtain}{user specification for which statistic/action they wish to obtain/perform}

\item{multiple}{clarify whether user is working with a single 'x3p' object or multiple in a tibble style dataframe}

\item{select_col}{if multiple = TRUE, location of the x3p objects}

\item{value}{user specified value for rescaling surface elements of an 'x3p' object}
}
\value{
Statistics regarding surface of 'x3p' object

Rescaled surface.matrix of 'x3p' object
}
\description{
This function takes a single or multiple x3p objects and allows
a user to obtain summary statistics regarding the surface.
}
\details{
This function also takes single or multiple x3p objects and allows
the user to rescale the surface elements.
}
