% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_to_tensor.R
\name{x3p_to_tensor}
\alias{x3p_to_tensor}
\title{Create Tensor from 'x3p' object}
\usage{
x3p_to_tensor(
  df,
  array_dim,
  dim1,
  dim2,
  dim3,
  dim4,
  multiple = FALSE,
  select_col
)
}
\arguments{
\item{df}{object of class 'x3p' or a tibble style dataframe containing 'x3p' objects}

\item{array_dim}{specify the tensor dimension}

\item{dim1}{dimension 1, x}

\item{dim2}{dimension 2, y}

\item{dim3}{dimension 3, z, surface values}

\item{dim4}{dimension 4, color scale}

\item{multiple}{clarify whether user is working with a single 'x3p' object or multiple in a tibble style dataframe}

\item{select_col}{if multiple = TRUE, location of the x3p objects}
}
\value{
A array/tensor with user specificed dimensions
}
\description{
This function takes a single 'x3p' object or multiple and creates a tensor used for training Deep Learning Algorithms
}
